pyproject.toml
[project]
name = "yolo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.2.1",
    "streamlit>=1.44.1",
    "ultralytics>=8.3.111",
]
